chore: see to review comments and refactor
