+++ package.json
@@ -126,7 +126,6 @@
     "ts-node": "10.9.2",
     "tw-animate-css": "1.4.0",
     "typescript": "5.9.3",
-    "vite": "8.0.13",
     "vite-tsconfig-paths": "6.1.1",
     "vitest": "4.1.6"
   }
