--- src/libs/icons/brandIcons.tsx
@@ -1,79 +1,0 @@
-import { createLucideCompatIcon } from './createLucideCompatIcon';
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Facebook = createLucideCompatIcon(
-  'facebook',
-  <path d="M18 2h-3a5 5 0 0 0-5 5v3H7v4h3v8h4v-8h3l1-4h-4V7a1 1 0 0 1 1-1h3z" />,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Github = createLucideCompatIcon(
-  'github',
-  <>
-    <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4" />
-    <path d="M9 18c-4.51 2-5-2-7-2" />
-  </>,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Gitlab = createLucideCompatIcon(
-  'gitlab',
-  <path d="m22 13.29-3.33-10a.42.42 0 0 0-.14-.18.38.38 0 0 0-.22-.11.39.39 0 0 0-.23.07.42.42 0 0 0-.14.18l-2.26 6.67H8.32L6.1 3.26a.42.42 0 0 0-.1-.18.38.38 0 0 0-.26-.08.39.39 0 0 0-.23.07.42.42 0 0 0-.14.18L2 13.29a.74.74 0 0 0 .27.83L12 21l9.69-6.88a.71.71 0 0 0 .31-.83Z" />,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Instagram = createLucideCompatIcon(
-  'instagram',
-  <>
-    <rect width="20" height="20" x="2" y="2" rx="5" ry="5" />
-    <path d="M16 11.37A4 4 0 1 1 12.63 8 4 4 0 0 1 16 11.37z" />
-    <line x1="17.5" x2="17.51" y1="6.5" y2="6.5" />
-  </>,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Linkedin = createLucideCompatIcon(
-  'linkedin',
-  <>
-    <path d="M16 8a6 6 0 0 1 6 6v7h-4v-7a2 2 0 0 0-2-2 2 2 0 0 0-2 2v7h-4v-7a6 6 0 0 1 6-6z" />
-    <rect width="4" height="12" x="2" y="9" />
-    <circle cx="4" cy="4" r="2" />
-  </>,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Slack = createLucideCompatIcon(
-  'slack',
-  <>
-    <rect width="3" height="8" x="13" y="2" rx="1.5" />
-    <path d="M19 8.5V10h1.5A1.5 1.5 0 1 0 19 8.5" />
-    <rect width="3" height="8" x="8" y="14" rx="1.5" />
-    <path d="M5 15.5V14H3.5A1.5 1.5 0 1 0 5 15.5" />
-    <rect width="8" height="3" x="14" y="13" rx="1.5" />
-    <path d="M15.5 19H14v1.5a1.5 1.5 0 1 0 1.5-1.5" />
-    <rect width="8" height="3" x="2" y="8" rx="1.5" />
-    <path d="M8.5 5H10V3.5A1.5 1.5 0 1 0 8.5 5" />
-  </>,
-);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Twitch = createLucideCompatIcon('twitch', <path d="M21 2H3v16h5v4l4-4h5l4-4V2zm-10 9V7m5 4V7" />);
-
-/** Lucide v0 brand mark (removed in lucide-react v1). */
-export const Youtube = createLucideCompatIcon(
-  'youtube',
-  <>
-    <path d="M2.5 17a24.12 24.12 0 0 1 0-10 2 2 0 0 1 1.4-1.4 49.56 49.56 0 0 1 16.2 0A2 2 0 0 1 21.5 7a24.12 24.12 0 0 1 0 10 2 2 0 0 1-1.4 1.4 49.55 49.55 0 0 1-16.2 0A2 2 0 0 1 2.5 17" />
-    <path d="m10 15 5-3-5-3z" />
-  </>,
-);
-
-/** Lucide v0 paths (redrawn in lucide-react v1). */
-export const CircleUserRound = createLucideCompatIcon(
-  'circle-user-round',
-  <>
-    <path d="M18 20a6 6 0 0 0-12 0" />
-    <circle cx="12" cy="10" r="4" />
-    <circle cx="12" cy="12" r="10" />
-  </>,
-);
